requirements.txt
dj-static==0.0.6
Django==1.8.5
django-appconf==1.0.1
django-compressor==1.5
djangorestframework==3.2.4
drf-nested-routers==0.10.0
six==1.10.0
static3==0.6.1
wheel==0.24.0
